+++ src/vector.c
@@ -8,24 +8,24 @@
 vector* vector_init()
 {
     vector *v = calloc(1, sizeof(vector));
-    v->vect = calloc(VECTOR_INIT, sizeof(uint32_t));
+    v->vect = calloc(VECTOR_INIT, sizeof(uint64_t));
     v->size = VECTOR_INIT;
     v->tos = 0;
 
     return v;
 }
 
-void vector_push_back(vector *v, uint32_t value)
+void vector_push_back(vector *v, uint64_t value)
 {
     if(v->tos == v->size)
     {
         v->size <<= 1;
-        v->vect = realloc(v->vect, sizeof(uint32_t)*v->size);
+        v->vect = realloc(v->vect, sizeof(uint64_t)*v->size);
     }
     v->vect[v->tos++] = value;
 }
 
-int vector_find(vector *v, uint32_t value)
+int vector_find(vector *v, uint64_t value)
 {
     for(int i=0; i < v->tos; i++)
         if(v->vect[i] == value)
